Add function to get dyld information on OSX
On OSX systems it's helpful to also get the list of dylib's loaded up by a process.
The get_process_maps call returns the memory for these regions as being in
'/private/var/db/dyld/dyld_shared_cache_x86_64h' instead of the dylib filename.
When trying to inspect the system ruby or python frameworks, we need to get the
actual dylib file instead (something like
/System/Library/Frameworks/Python.framework/Versions/2.7/Python).

This adds a function
to get a list of all the loaded dylibs from a target process, as well as the
correct virtual memory addresses for the __TEXT/__DATA sections.